Move client-env file
