+++ packages/client/src/lib/logger-lib.ts
@@ -1,5 +1,5 @@
-import ENV from "lib/client-env";
 import { createLogger } from "redux-logger";
+import ENV from "tools/client-env";
 
 /** ===========================================================================
  * Redux Logger Setup
