+++ packages/client/src/modules/app/epics.ts
@@ -9,7 +9,6 @@ import {
 } from "@anthem/utils";
 import { graphqlSelector } from "graphql/queries";
 import Analytics from "lib/analytics-lib";
-import ENV from "lib/client-env";
 import { EpicSignature } from "modules/root";
 import { i18nSelector } from "modules/settings/selectors";
 import { combineEpics } from "redux-observable";
@@ -24,6 +23,7 @@ import {
   take,
   tap,
 } from "rxjs/operators";
+import ENV from "tools/client-env";
 import {
   getQueryParamsFromUrl,
   initializeNetwork,
